website overhaul
I've changed the entire color of the website the fonts and the backgroud
image in the left portion of the website. The remaining portion that
needs to be edited is the design of the button, the credits and
'already have an account' caption.

Also need fix on the input fields default color, default color of the
broder of the input fields becomes red due to input:valid css. Also
include additional validations in the input fields.